VSM-927: DeviceData::int maxFps -> DeviceData::float fps.
<==== Mercurial data ====>
Branch: webcam_plugin

+++ plugins/usb_cam/src/ffmpeg/codec_parameters.h
@@ -14,12 +14,12 @@ namespace ffmpeg {
 struct CodecParameters
 {
     AVCodecID codecID;
-    int fps;
+    float fps;
     int bitrate;
     int width;
     int height;
 
-    CodecParameters(AVCodecID codecID, int fps, int bitrate, int width, int height):
+    CodecParameters(AVCodecID codecID, float fps, int bitrate, int width, int height):
         codecID(codecID),
         fps(fps),
         bitrate(bitrate),
